Home refactoring and google connect link #8
