+++ frontend/src/components/Home.tsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from 'react';
 import { BackendModel } from '../BackendModel';
+import CreateWorkflow from './CreateWorkflow';
 
 function Home() {
 
     const [workflows, setWorkflows] = useState<BackendModel.Workflow[]>([]);
     const [lastCreatedWorkflow, setLastCreatedWorkflow] = useState<BackendModel.Workflow | undefined>(undefined);
-    const [newWorkflowName, setNewWorkflowName] = useState<string>("");
 
     useEffect(() => {
         fetch("http://localhost:3000/workflows")
@@ -13,20 +13,6 @@ function Home() {
             .then(data => setWorkflows(data))
             .catch(error => console.error(error));
     }, [lastCreatedWorkflow]);
-
-    const createWorkflow = () => {
-        if (newWorkflowName !== "") {
-            const requestOptions = {
-                method: 'POST',
-                headers: {'Content-Type': 'application/json'},
-                body: JSON.stringify({name: newWorkflowName, nodes: [], edges: []})
-            };
-            fetch("http://localhost:3000/workflows/new", requestOptions)
-                .then(response => response.json())
-                .then(data => setLastCreatedWorkflow(data))
-                .catch(error => console.error(error));
-        }
-    }
 
     const deleteWorkflow = (workflow: BackendModel.Workflow) => {
         const requestOptions = {method: 'DELETE'};
@@ -39,8 +25,10 @@ function Home() {
     return (
         <div>
             <div>
-                <input type="text" placeholder="Workflow name" value={newWorkflowName} onChange={(e) => setNewWorkflowName(e.target.value)} />
-                <button onClick={createWorkflow}>Create a new workflow</button>
+                <CreateWorkflow setLastCreatedWorkflow={setLastCreatedWorkflow}/>
+                <a href="http://localhost:3000/auth/google/login" target="_blank" rel="noopener noreferrer">
+                    Connect a Google account
+                </a>
             </div>
             <ul>
                 {workflows.map(workflow => <li key={workflow._id}><a href={`/workflow/${workflow.name}/${workflow._id}`}>{workflow.name}</a><button onClick={() => deleteWorkflow(workflow)}>X</button></li>)}
